Perform an estimation of the covariance with local linear smoothers.

This function performs the estimation of the covariance of a set of curves using local linear smoothers where the bandwidth is estimated using the methodology from Golovkine et al. (2021).

covariance_ll(
  curves,
  grid = seq(0, 1, length.out = 101),
  grid_param = seq(0.1, 0.9, by = 0.1),
  grid_bandwidth = NULL,
  center = TRUE,
  delta_f = NULL,
  n_obs_min = 2,
  kernel_name = "epanechnikov"
)

Arguments

curves

List, where each element represents a curve. Each curve have to be defined as a list with two entries:

$t Sampling points

$x Observed points.

grid

Vector (default = seq(0, 1, length.out = 101)), sampling points at which estimate the curves.

grid_param

Vector (default = seq(0.1, 0.9, by = 0.1)), the sampling points at which we estimate the parameters.

grid_bandwidth

Vector (default = NULL), grid of bandwidths.

center

Boolean (default = TRUE), center the data?

delta_f

Function (default = NULL), function to determine the delta.

n_obs_min

Integer (default = 2), minimum number of observation for the smoothing.

kernel_name

String (default = 'epanechnikov'), the kernel used for the estimation:

epanechnikov

uniform

biweight

Value

List of with three entries:

$parameters Estimated parameters.

$bandwidths_mat Estimated bandwidths matrix.

$covariance Estimated covariance.

References

Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive estimation of irregular mean and covariance functions.