Perform an estimation of the bandwidth for the estimation of the mean.

This function performs an estimation of the bandwidth for a univariate kernel regression estimator defined over continuous data.

estimate_bandwidth_mean(
  curves,
  params,
  sigma = 0,
  grid_bandwidth = lseq(0.001, 0.1, length.out = 101),
  n_obs_min = 2,
  kernel_name = "epanechnikov"
)

Arguments

curves

List, where each element represents a curve. Each curve have to be defined as a list with two entries:

$t Sampling points

$x Observed points.

params

List, estimation of the different parameters:

$point Time point where the smoothing has been done.

$H Estimated regularity.

$L Estimated constant.

$var Estimated variance.

sigma

Numeric, estimation of the std of the noise \(\sigma\).

grid_bandwidth

Vector (default = lseq(0.001, 0.1, length.out = 101)), grid of bandwidths.

n_obs_min

Integer (default = 2), minimum number of points in the neighborhood to keep the curve in the estimation.

kernel_name

String (default = 'epanechnikov'), the kernel used for the estimation:

epanechnikov

uniform

biweight

Value

List, estimation of the bandwidth.

References

Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive estimation of irregular mean and covariance functions.

See also