Perform an estimation of the mean with local linear smoothers.

This function performs the estimation of the mean of a set of curves using local linear smoothers where the bandwidth is estimated using the methodology from Golovkine et al. (2021).

mean_ll(
  data,
  U = seq(0, 1, length.out = 101),
  t0_list = 0.5,
  grid = NULL,
  nb_obs_minimal = 2,
  kernel = "epanechnikov"
)

Arguments

data
A list, where each element represents a curve. Each curve have to be defined as a list with two entries:

$t The sampling points

$x The observed points.

U
Vector, sampling points at which estimate the curves.

t0_list
Vector, the sampling points at which we estimate the parameters.

grid
Vector, default=NULL. A grid of bandwidths.

nb_obs_minimal
Integer, minimum number of observation for the smoothing.

kernel
Character string, the kernel used for the estimation:

epanechnikov (default)

uniform

biweight

Value

A list of with two entries:

$parameter The estimated parameters.

$mu The estimated mean.

References

Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive estimation of irregular mean and covariance functions.