Perform a non-parametric smoothing of a set of curves for mean estimation.

This function performs a non-parametric smoothing of a set of curves using the Nadaraya-Watson estimator.

smooth_curves_mean(
  curves,
  grid = NULL,
  grid_param = c(0.25, 0.5, 0.75),
  grid_bandwidth = NULL,
  delta_f = NULL,
  kernel_name = "epanechnikov",
  n_obs_min = 2
)

Arguments

curves

List, where each element represents a curve. Each curve have to be defined as a list with two entries:

$t Sampling points.

$x Observed points.

grid

Vector (default = NULL), sampling points at which estimate the curves. If NULL, the sampling points for the estimation are the same than the observed ones.

grid_param

Vector (default = c(0.25, 0.5, 0.75)), sampling points at which we estimate the parameters.

grid_bandwidth

Vector (default = NULL), grid of bandwidths.

delta_f

Function (default = NULL), function to determine the delta.

kernel_name

String (default = 'epanechnikov'), the kernel used for the estimation:

epanechnikov

uniform

biweight

n_obs_min

Integer (default = 2), minimum number of observation for the smoothing.

Value

A list, which contains two elements. The first one is a list which contains the estimated parameters:

sigma Estimation of the standard deviation of the noise.

variance Estimation of the variance of the process.

H0 Estimation of \(H_0\).

L0 Estimation of \(L_0\).

b Estimation of the bandwidth.

The second one is another list which contains the estimation of the curves:

$t Sampling points.

$x Estimated points.

References

Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive estimation of irregular mean and covariance functions.